pyproject.toml
[tool.poetry]
name = "boers"
version = "0.1.0"
description = "Prueba tecnica"
authors = ["nleea <neldecas12@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
asgiref = "3.8.1"
aspectlib = "2.0.0"
attrs = "24.2.0"
certifi = "2024.8.30"
cffi = "1.17.0"
charset-normalizer = "3.4.0"
coreapi = "2.3.3"
coreschema = "0.0.4"
cryptography = "43.0.0"
django = "5.1.4"
django-cors-headers = "4.4.0"
django-environ = "0.11.2"
djangorestframework = "3.15.2"
djangorestframework-simplejwt = "5.3.1"
drf-spectacular = "0.27.2"
environ = "1.0"
fields = "5.0.0"
gunicorn = "23.0.0"
idna = "3.10"
inflection = "0.5.1"
itypes = "1.2.0"
jinja2 = "3.1.4"
jsonschema = "4.23.0"
jsonschema-specifications = "2024.10.1"
markupsafe = "3.0.1"
openapi-codec = "1.3.2"
packaging = "24.1"
psycopg = "3.2.1"
pycparser = "2.22"
pyjwt = "2.9.0"
pytz = "2024.2"
pyyaml = "6.0.2"
referencing = "0.35.1"
requests = "2.32.3"
rpds-py = "0.20.0"
setuptools = "75.2.0"
simplejson = "3.19.3"
sqlparse = "0.5.1"
typing-extensions = "4.12.2"
tzdata = "2024.1"
uritemplate = "4.1.1"
urllib3 = "2.2.3"
pymysql = "^1.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
